Allow links to contain ID fragments.
Currently, the on click function contains a conditional checking if the href contains a `#` and blocks that with `e.preventDefault()`. This is problematic for people wishing to use the menu to link to anchors with a name or id fragment, for example: `<a href="examplepage.html#contrivedfragment">YAS KWEEN</a>`

Adjusting the regex just a touch unblocks this to allow for name or id fragments while still blocking links with **only** a `#`.